Simplify Makefile and add linking of library
Instead of setting LD_LIBRARY_PATH each time I added a flag to load it at compile, I am unsure of the details here. Also the old format of the make file directives was unclear. I simplified it for now, but may need to make it more complex in the future.