Implement new format for general couplings
This format allows users to define a coupling energy function and pass it as a parameter to the constructor.
The form of the energy function has position, x, y, z, and, for now, a spin attribute for each 'node', or 'grain', or 'particle' etc...
Additionally this format for couplings implements a common coupling, square with neighbors being up, down, left, and, right. It allows for periodic or non-periodic boundary conditions.
Removes couple and decouple functions as they somewhat repeat the functionality of the overloaded operator.