Create an experiment to test ergodicity
The results of this indicate the number of reachable states by RECA, for large L, is the fraction 1/(N^L) of the total number of states. Where N is the number of spins a node can have and L is the number of nodes in the state, thus 2L including the replica.