Simplify Bonds
Bonds is now a three tier structure. The pure virtual class bonds
contains the core of what Bonds are and is the first tier. The
second tier is structural, how are site/nodes connected?
StaticCouplings2D is the only class fulfilling this role currently.
The third tier is defining values for Bonds, which in physical
systems may represent the energy of the bond. ABonds and JBonds
represent this level. Additionally, it is helpful to have
everything relating to a physical model together. JJA and Ising
models have been implemented using the new structure.